+++ src/types/counter.ts
@@ -2,7 +2,7 @@ interface Counter {
     [key: string]: number | undefined;
     uppercase?: number;
     lowercase?: number;
-    numbers?: number;
+    number?: number;
     special?: number;
 };